Add a second SV to non-devnet configuration (#6299)
* Add a second SV to non-devnet configuration

fixes #6295

This adds a second canton-foundation-2 SV to replace Daml Hub for
now. The preflight checks are extended to cover this.

In non-devnet this SV does not have its own sequencer/mediator just
like external SVs.

I didn't preserve the option to run with only a single SV. It seems
unlikely that we test this anywhere for now and in two weeks we can
hopefully switch back to a single SV anyway so the additional
complexity does not seem worth it.