Switch testnet-preview to automatic deploy (#6383)
fixes #6379

As usual for these types of changes unfortunately not really testable
without merging. The CI jobs work though so hopefully this works out
fine, I'll keep an eye on it.

Note that the mark-for-testnet stuff still happens as part of devnet
for now. We may want to switch that at some point but not for now.