Support bootstrapping a cluster from an ACS dump (#6401)
fixes #6396

I'm not really sure how to write a good test for this given that this
isn't used on any cluster atm. Obviously tested it manually.

My plan is to make testnet-preview always deploy with the most recent
ACS and identity backups which should allow us to test some stuff
there. (although I'm still not sure what exactly we would test for
ACS imports, for identity imports it's relatively clear)