Enable SV app to initialize its participant from identities dump (#6417)
Part of #6178

This enables SVs to migrate their identities across network resets, without messy bootstrap scripts. The next PR for this issue will be about enabling non-SV validators to do the same.

One sad learning: There doesn't seem to be a way to tell a participant that was started with `auto-init=false` that it should just generate a new ID after all. I.e., we have to know if we have a dump when we start the participant, and tweak the participant config accordingly.