Support bootstrapping a cluster from participant identity backups (#6428)
* Support bootstrapping a cluster from participant identity backups

part of #6405

The only thing missing to fix it is for the validator app to properly
do its job so I can set the participant init to false. I added some
dummy cdoe to check that I can read the dump even there.

It does work for the SV app.