Bootstrap testnet and testnet-preview from preceeding cluster (#6471)
I unfortunately have no reasonable idea how to test this without
merging it and trying it out.

I did run `cncluster ci_apply` locally against a cluster and that
worked :shrug:

fixes #6461

Note that this will not work for testnet automatically, The snapshots
don't have the right format atm. We will need to manually upload and
reshuffle them in the GCP bucket once.

On a related note, I think it also eventually makes more sense to
bootstrap testnet-preview from testnet as opposed to
testnet-preview. But that doesn't work for the same reason atm.