Don't include version in gcp bucket file names (#6479)
* Don't include version in gcp bucket file names

Figuring out the right version is fairly annoying and we don't gain
much by including it in the filename.

This PR instead includes it in the dump (optional for backwards
compat).

The restore logic is not backwards compatible wrt to the file
location. We already broke this heavily so we jut have to do one
manual copy once which isn't that terrible.