Switch splitwell to JSON API (#6741)
And thereby get rid of envoy for everything user facing.

We still have an envoy around the splitwell backend for now, will work
on killing that next.

I kept the JSON API on port 7575 in our cluster for now to avoid breaking
ingress settings. I think we should probably switch that though. If
you agree, I'll create a tech debt