Quick fix for busy looping triggers (#6858)
Fixes #6810 via an easy workaround, so we have high confidence that the issue will be solved in the next round of deployments.

I created #6856 to follow up with something less ducktape-y.

I deployed an instance of this to scratcha. I wasn't able to get a trigger to busy loop there so far. (Preflights are passing.)