Extend app manager with Oauth2 Authorization server (#6842)
This adds enough of an OAuth2 authorization server to allow a
third-party application to get a token that can then be used to talk
to the ledger API.

This follows an OAuth2 authorization code grant flow. Note that you
don't see an endpoint for /oauth2/authorize. That's not actually an
API call but a redirect and in our case that redirect goes to the app
manager frontend rather than the backend. There is the call the
frontend makes to the backend though in /authorize. A few PRs in the
future (or in the past because the code already exists in the PR that
I'm splitting up?  timelines are confusing) I'll add the frontend for
that.

part of #6839