Add JSON API proxy to app manager to add CORS headers (#6843)
Talking to user's participants is a CORS request so we need a JSON API
that exposes CORS headers for this to work.

This is fine from a security perspective. The two main things CORS restrictions
protect against are:

1. Exploiting existing cookies to make requests from a different webpage.
2. Exploiting access to apps from user’s browsers that don't require
   authentication (e.g. because they rely on network-level controls).

The JSON API does not use cookies and it does require auth. Ofc if you
compromise the JS in a page that has a token you can still make a
request but whether that's a CORS request or not doesn't matter at
that point.

We should probably upstream this into Canton at some point but for now
this seems easy enough to manage that just baking this into the app
manager is fine. Having the proxy might also come in handy for other
things (e.g. if we want to experiment with restricted ledger api
access this might make for very cheap PoCs).