Fail hard if app versions mismatch (#6895)
Modifies the version compatibility check we do on the client side of our apps to fail hard instead of just logging an error message. This is a first step towards ensuring that issues like #6769 do not recur and should also prevent hard-to-debug errors due to partners running different code versions.