Retry on PERMISSION_DENIED (#7058)
see
https://support.digitalasset.com/s/case/5004x00000YrpLeAAJ/svapp-validatorapp-and-scanapp-stop-working-after-a-while-until-restarted

We don't have any integration tests for token refresh (we do have unit
tests for the refresh logic) atm unfortunately so this is also not
tested. The logs clearly show that we get PERMISSION_DENIED though so
confidence in this fixing (at least part of) the issue is high.