Add per-app users to app manager and auto-redirects (#7104)
* Add per-app users to app manager and auto-redirects

This adds per-app users to the app manager auth flows so we can
enforce restricted ledger API access (side note: we may want to
exploit the fact that we already proxy the JSON API for PoCs of those
restrictions). Currently they are still allocated in the default IDP.

It also adds automatic redirects in two places:

1. After clicking launch, the splitwell UI automatically triggers the
auth flow.
2. The /authorize UI endpoint automatically checks for authorization and
if so redirects back automatically. Note that the endpoint is still
called /authorize not /check-authorize or something like that because
that's the standard name in the OAuth2 flow.

That gets us to a single click app-launch.

part of #7092