Approve individual app release configurations (#7206)
* Approve individual app release configurations

This changes the install flow to not do anything other than cache the
configs. Instead there is a separate step to approve individual
release configurations.

Upgrading then becomes relatively straightforward: Discover a new
config and update the list of unapproved configurations. That's not
yet implemented in this PR though.

We may want to consider making the initial install special to get a
sightly nicer UX but for now this is much easier from an
implementation perspective and it seems a bit too early to optimize
the UX around this.

part of #7089