Identify release configs through hash of canonical JSON (#7227)
Much more reasonable than .hashCode which definitely does not do what
we  want.

I ended up reusing the implementation from metering. This doesn't feel
great given that it doesn't seem placed in a location that makes it
seem like it is intended for reuse. On the other hand, duplicating it
also seems dumb. Maybe want to consider upstreaming something to
Canton to move it.

part of #7089