Add JWT auth to app manager (#7265)
This splits the app manager API in three:

1. The admin API which can only be accessed by the validator
operator. This is required for app registration and installation mainly.
2. The per-user API that can be accessed by any user on the
validator. This is mainly for app launch.
3. The public API, things like the JWKS endpoint that has no auth at
all.

95% of this PR is just shuffling code around. The one actual change is
that the UI now accepts the provider user id as a separate field
instead of using the logged in user. We probably want to change that
to even allocate the user at some point (which would completely
subsume the app-instances config) but not in this PR.

I've followed the approach from the wallet to have per user client references.

fixes #7258