Add convenient way to mark the next `cidaily` for testnet (#7296)
To move the `testnet-next` tag to the commit of the next `cidaily` deployment we can now run `scripts/unmark-for-testnet.sh`. This will remove that tag. The next `cidaily` deployment will then recreate the tag. Note that if the next `cidaily` deployment fails, the next `testnet` deployment will too (for lack of a tag).

Since we plan to skip the next testnet release anyway (#7254), I'll test this manually this week after merging (don't know of a better testing strategy sadly).

Fixes #6935