Only wait for specific k8s pods in CI health checks even when sync_with_cluster_version is true (#7847)
Fixes #7843 by only waiting for the pods in the sv-1 to sv-4 namespace for the non-SV preflight check and for the pods in the sv namespace for the SV preflight check. This makes the behaviour of the cncluster wait step (during the preflight check job) with sync_with_cluster_version set to true the same as when it is false.