Store original payload value for explicit disclosure (#7960)
* Store original payload value for explicit disclosure

Converting to the Java codegen type is inherently lossy. We don't know
if the original payload was missing optional fields or if they were
just empty. However, for explicit disclosure we do need to have the
original payload available. Otherwise, the validation for
authenticated contract ids in Canton will fail.

Therefore, we now store both the original payload and the high-level
Java version. There is some storage duplication here that we could get
rid off. For now, I'd rather minimize the changes here and accept that
duplication. Killing it later is as easy as killing it now.

fixes #7955