Let pulumi install plugins (#8564)
This is done because pulumi needs all the versions of the plugins used by the current stack setup and the one that was previously deployed. This in turn leads to requiring multiple versions of the same plugin, and it's just not worth it keeping a list of all the plugins.