Add a section on documenting issue investigations (#9381)
* Add a section on documenting issue investigations

This is the result of a discussion between Nicu, Simon, Parth and
myself. We don't seem to have a clear answer for "this is the query
that we always use that needs to have a dashboard" yet so for now
let's start documenting it which seems benefitial at a minimum for
knowledge sharing and hopefully also to identify those queries.

Will post a message in slack to make everyone aware assuming nobody protests.