Upgrade to Canton ad-hoc snapshot to consume sequencer counter fix (#9494)
fixes #9387

This is an ad-snapshot built from
https://github.com/DACH-NY/canton/pull/16618.

I wasted way too much time to upgrade to a commit on `main` but
everything is broken so this won't work. This should at least gives
fewer issues in the meantime and unblock a devnet upgrade (the
sequencer counter issue seems way too frequent to upgrade with that).