Make trace-id allocation for LedgerApi calls good enough (#9758)
Includes:
- allocating and logging a fresh trace-id for the server-side call in the
  `ApiClientRequestLogger`
- logging missing and empty caller trace-ids in the `ApiClientRequestLogger`
- adding caller trace-id propagation to calls that already have an implicit
  `TraceContext` parameter

Excludes: adding trace-id propagation to getUser and like calls, as they are
used extensively within (variants of) `retry`. This is difficult to do, as
our retry infrastructure is missing appropriate trace-id management within
itself. Tracking ticket for improving this: #9754

It looks though that tracing for LedgerApi calls is good enough with this change,
which thus fixes #9713