Checkout version before restoring nix cache (#9780)
This should hopefully give us a decent chance of actually getting
cache hits. I did bump the resource class as well though, especially
on testnet there is a realistic chance the cache gets expired. We
could in theory only bump it there but didn't sem woth the complexity
for now.