Also grant featured app rewards for transfers with provider == sender (#10974)
Before this change we had an extra restriction in place to not issue featured
app rewards for CC transfers with sender == provider. The idea was to avoid
a featured app provider being accused of wash trading when they just sent
CC to somebody. However, they can also just not pass the featured app right
into the transfer when they want to avoid receiving non-justified featured
app rewards.

Moreover, explaining the restriction is cumbersome, and working around the
restriction is trivial (just allocate another party on the same node as the
provider party and use that one as an intermediary).

Thus I believe we're better off removing the restriction.