Nuke validator balance top-up preflight (#11064)
It already doesn't work reliably atm. On testnet there is no chance it
will work reliably because we bumped the round 0 duration to 24h see

For 24h round0 it is also sort of unfixable. We have to accept that we
don't get coin before. And if we can tolerate 24h we can probably
tolerate a whole week.