Set (starting) domain fees parameters for all clusters. (#11356)
Fixes #11287

Also fixes #11624 which was flaking often on this PR.

Based on calculations in #11299 (comment)

I also needed to tweak a few additional parameters to get this (further) through CI. Mainly the base rate burst amount is doubled to support validator onboarding. The burst window is also doubled correspondingly so that the rate remains the same.

Topups are now enabled for the self-hosted runbook tests and the polling interval for the topup trigger is reduced for these tests to allow the initial topups to take place quickly enough (and prevent the tests from failing due to lack of traffic).