Include cometbft validator index in list_sv_cometbft_addresses (#11962)
Cometbft often prints the index in addition to the key which can be
easier to compare. However, the script so far reordered everything
more or less randomly which it easy to get tripped up by that.

Now the order is preserved and we explicitly include the index so you
don't have to count array elements.