Support `global.canton.network.digitalasset.com` in infra (#12188)
Part of #11913 

This PR adds support for `global.canton.network.digitalasset.com` in:
DNS entries
certificates
istio gateway
auth0 redirects

(all four of them support both domains for now. That's probably best to keep for dev at least, and we can revisit it for prod at a later point)

With this in place, any VirtualService that is configured with a global.canton.network.digitalasset.com hostname works (but in this PR I don't yet modify any VirtualServices).