Add check for dars.lock immutability (#12746)
* Add check for dars.lock immutability

This is a subset of #12559 to catch obviously broken changes like the
API docs change I messed up.

It has some overlap with the check in the script for conflicting
hashes but that one can be harmless if we make multiple changes within
one release becasue we only bump the version number once. This one on
the other hand should really always be a hard failure.