Add option to disable update hist ingestion from participant begin (#12782)
Backfilling from participant begin is way too slow on devnet atm and
the participant ends up overloaded so we don't want to do that.

This is primarily an issue for the SV app as the DSO party has orders
of magnitude more history than the rest.

I expect we only enable this for the SV app. I added it to the
validator app and scan for completeness but only as a helm option for
the validator and not for scan to avoid anyone accidentally enabling
it there (it should not make a difference since we already have
history but why risk it).