Make tap tx log assertions more robust (#13311)
* Make tap tx log assertions more robust

fixes #13195

My theory is that what is happening is that there is earlier tap (the
logs confirm that there is one 10-20min before) of the same amount
which is still in there (which is expected in the upgrade tests) and
this means we end up filtering it out. The reason why it's flaky is
that depending on how many reward collection or whatever else happens
in between the old tap might get pushed out of the tx log already.

Relying on tx dates seems more robust anyway so that's what this PR
switches to.

I also increased log output so in the (likely) case that I'm
misreading it and it does not fix the issue we at least have a
reasonable chance to debug it.