Add retries to all CI-run `pulumi up` jobs except the `cimain` ones (#13397)
Attempted workaround for issues such as #11746 #13304 #12575

"Rerunning" is what we're doing on these failures anyway. If the operator was managing those clusters it would also be retrying failed ups, so this has some consistency with that.

This PR also rewires Slack notifications a bit and adds Slack notification on failed deployments that we're retrying on.