Fix "busy automation" alert to account for parties and polling-only triggers (#15086)
Ended up splitting it into two largely identical alerts. There are triggers that are both polling and task-based and I couldn't figure out a nicer way to account for this.

Also more generally: I decided to timebox making this nicer and also cleaning up the associated dashboard. This Fixes #14833 and this is the only thing that counts for now :)