Fix conversion between string and long offset (#15725)
The converters treat an empty string and 0 as the same value. So if we
write offset 0 to the store we actually write an empty string. But we
read it as `None` and then some tests failed. I saw it flake once with
an error and the unit test here fails reliably without the fix.

[ci]

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>