Revert "Add amulet name and acronym on external rules (#15788)" (#15826)
This reverts commit 3c5c02f4103e3e8892fbc61afe7bb04aca8effd3.

Adding the human-readable name gives a false sense of security, as
what really matters is the party-id of the `dso` party. Moreover,
having the identifier on-ledger complicates the factory implementations,
and in particular, will require to also push the identifier onto
the `AmuletRules` themselves, as they can't read it from the
`ExternalPartyAmuletRules` due to circular dependencies.