Don't run integration tests on minor release branches (#15953)
See #15944 and #15952 for preceding discussion.

It doesn't seem like there is anything relevant that we could break in a `release-line-x.y.z` branch that we would need an integration test to catch. What we do get is all the flakes that we didn't backport fixes for, plus CI costs.

This PR embodies the semantics: Make sure things work before cutting a release. Once you did, we don't check the release branch anymore for anything but deployments-related stuff (except if it's a major release).