First cut at docs restructuring (#17090)
* First cut at docs restructuring

[ci]

This gets some basic high-level structure in place. I did end up
option for the todo annotations but only display them in live
preview (different from what we had before where they were also
included in the prod build).

I mostly avoided empty sections and just added todos but I did add a
few where it seemed easier to have them than not have them.

I did not make any attempts at filling in the missing content.

Sphinx fails on pages not included in the ToC so we don't need to
check for that.

I didn't aggressively move files around in the repo as that breaks
links but I also didn't overoptimize for keeping them stable. It's
easy enough to find stuff in the toc if someone ends up with a broken
link.

Probably easiest to review with livepreview locally.

* Add headers

* Add missing pieces discovered during review

[ci]

* make the name checker happy

[ci]

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Simon Meier <simon@digitalasset.com>