Add validator hardware requirements (#17138)
* Add validator hardware requirements

fixes #17116

The three categories are the ones Wayne asked for. The actual values
are somewhat random. The local laptop one seems to roughly correspond
to what I get when I run a docker compose validator locally. The next
one is roughly what I get for validator1/splitwell. And lastly, the
last one roughly corresponds to the registry app.

[ci]

* links are good

[ci]

* add headers

[ci]

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>