Improve documentation around network parameters (#17216)
* Improve documentation around network parameters

[ci]

This adds a section very close to the top describing the 3 (or 4 for
helm) parameters that define your connection to the network and
includes a link to the GSF page that includes the migration id. Those
are the ones we've seen people get wrong the most often so making this
more prominent seems important.

fixes #17139

* review feedback

[ci]

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>