Remove references to DA VPN from the docs (#17385)
* Remove references to DA VPN from the docs

[static]

I've dropped the VPN mention completely from the helm docs. I kept it
in the compose docs which just mention that you need to have _a_
whitelisted VPN or static egress but not that it should be the DA
one. We may also want to drop that but let's come up with a good plan
for doing so first.

* Update docs/src/sv_operator/sv_helm.rst

[static]

Co-authored-by: Martin Florian <martin.florian@digitalasset.com>

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Martin Florian <martin.florian@digitalasset.com>