Skip validator bootstrap script if no DARs are specified (#17574)
[ci]

fixes #17573

Tested locally in both configurations.

I think long term moving the DAR upload in the validator app may still
make more sense but Canton is planning to add similar functionality to
the participant itself and I haven't seen anyone use this other than
us in tests so not bothering with that.

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>