Fix acknowledgement catchup dashboard (#17810)
increase is intended for counters where you can reasonable assume that
if it went down you wanna start counting again from 0. This makes no
sense for acknowledgements.

Arguably Canton should also fix their metrics to not jump back in time
on a cometbft replay (the store does seem to take the maximum). But
it's actually kind of useful for identifying cometbft issues so not
requesting that for now.

[static]

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>