Remove java codegen workaround from token-standard (#17993)
fixes #17962

[static]

Now that the code is in splice I see no reason to not just tweak the
daml.yaml files. Other workarounds are definitely possible but this
seems perfectly adequate for our setup. External users can just
specify the token std as a separate argument to generate code from it
which also seems fine.

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>