Ignore `Sequencer shutting down` cluster logs (#18661)
Common on DevNet these days.

Judging by the timing and the fact that they are connected to different sequencers at different times this seems to be simply noise related to sequencer restarts.

Not sure why we didn't get noise about this previously, but also don't see good reason to not ignore.