Fix an update topology compat check instructions (#18952)
The real issue is the missing TimeQuery.Range(None, None). That can
result in export higher serials without exporting the prior ones which
then obviously fails. Our prod code seems to do this correctly
already.

The rest is just updating to the 3.3 APIs. Note that this goes a step
too far technically. If you want to test compat 3.2 -> 3.3 the export
still needs to use the old stuff. But at this point just updating
everything seems to make more sense to make the instructions as close
to what would work for 3.3 -> 3.4 as possible.

[static]

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>