Fix scan retries (#19255)
[static]

Somehow was misremembering that RetryFor.Automation retries forever,
it doesn't. We have RetryFor.LongRunningAutomation but it also doesn't
do what we want, it's for automation that fails infrequently,
e.g. update stream ingestion but if you get repeated failures it still
gives up.

We could add a "retry forever" but doesn't seem worth
it. WaitingOnInitDependency is long enough that it works well in
practice. Even RetryFor.Automation was long enough that it fixed the
original issue.

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>