Add docs for recovering coin balance of an external party (#19556)
* Add docs for recovering coin balance of an external party

fixes #19451

[static]

See #19451 for why I'm recommending a fresh participant. Just seems
way too dangerous to try to reuse an existing one and then have it
blow up. That also means submitting the topology transaction through
that participant is fine. Worst case it blows up and you just try
again on another node which is the same procedure we have for
non-external parties.

All seems way too complex but don't really see a great way to simplify
it all that much.

* Apply suggestions from code review

Co-authored-by: Nicu Reut <nicu.reut@digitalasset.com>

* review

[static]

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Nicu Reut <nicu.reut@digitalasset.com>