Fix wallet txlog parsing for two-step token standard transfers (#19614)
* Fix wallet txlog parsing for two-step token standard transfers

This ensures that both sender and receiver parse all changes to a
transfer instruction and changes to balances are merged into the same
entry.

I opted for creating too many entries instead of too little. In
particular, it creates tx log entries for things that do not directly
change the balance, e.g., the receiver has an entry for the offer
being created. I still abused balance change and tx log entries for
that, turning it into notifications just adds a ton of duplication.

The amounts might be slightly unintuitive as the wallet only tracks
unlocked amounts so creating the offer is the part that reduces your
balance and accepting it actually increases your balance since the
remainder gets transferred back. But for now this is the best we can
do, I'm definitely not going to start exposing locked amulets in the
wallet UI.

Will do the UI after I fix the scan tx log entries.

part of #19607

[ci]

* Update apps/wallet/src/main/scala/org/lfdecentralizedtrust/splice/wallet/store/UserWalletTxLogParser.scala

Co-authored-by: Robert Autenrieth <31539813+rautenrieth-da@users.noreply.github.com>

* more comments

[ci]

* Fix interface id parsing

[ci]

* my formatting is perfect, shutup CI

[ci]

* fix

[ci]

* .

[ci]

* header

[ci]

* stop UI from crashing, will make it pretty in a follow-up

[ci]

---------

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Robert Autenrieth <31539813+rautenrieth-da@users.noreply.github.com>