Fix scan txlog parsing for TransferCommand_Send (#19644)
Apparently we forgot to recurse which means while we created the entry
that is used by the transfer command status endpoint, we missed the
actual transfer entry.

In theory this needs backfilling to fix the history. In practice, the
impact is fairly small. It's limited to the scan history and old
round-based balance endpoints. The latter has been deprecated forever
and the tx history is replaced by the update history outside the scan
UI. The amount of actual transfers initiated by external parties on
mainnet is also fairly small.

But more importantly the txlog backfilling stuff is also not yet
merged so not much we can do either way for now.

[ci]

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>