Relax tx history assertion in sv reonboarding tests (#19648)
[static]

I briefly looked at factoring out the duplication with the other test
but it's actually slightly diffeernt so going for this for now.

I didn't test it but the same fix worked for the other test and it's
pretty obviously only a less strict assertion so it shouldn't make
anything worse.

fixes https://github.com/DACH-NY/cn-test-failures/issues/4206

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>