Switch to new ACS export/import endpoints
[ci]

This breaks cross-version compatibility in a few ways:

1. Sponsor SV needs to be on same version as the candidate.
2. For recover from key scan needs to be on the same version as the
validator.
3. A HSM export needs to happen on the same version as the import.

1 and 2 don't seem problematic to me. 3 is a bit more problematic. In
particular, we need to avoid merging this while we might still want to
cut a new 3.x release that we upgrade testnet or mainnet to directly
and even afterwards there may be stragglers. We can either tell people
they need to upgrade to 0.4.1 first to complete the HDM which sounds
fairly reasonable or we opt for the safer option of waiting for 30
days after the HDM is complete because then they would be pruned away
anyway.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>