Fix migrations for SV validators
Fixes https://github.com/DACH-NY/cn-test-failures/issues/4829

Bug was introduced in https://github.com/hyperledger-labs/splice/pull/1297

This was not caught in our `DecentralizedSynchronizerMigrationIntegrationTest` because we don't migrate SV2's validator there.
Extending the (already slow and complex) test to do so doesn't seem worthwhile to me though; also as we plan to phase out synchronizer migrations entirely in the not-too-long term.

[ci]

Signed-off-by: Martin Florian <martin.florian@digitalasset.com>