[ci] Fix issue where multiple caches could've been created in tests
The root cause is that TrieMap.getOrUpdate guarantees that it returns the same element all the time but the function passed to generate the element can be called multiple times.
This in turn was creating multiple caches, that wouldn't have been an issue but the underlying gauges would have never been GCd because they are referenced by the metrics infra to read the values.
The new implementation is safe because the gauges are registered only as a final step when creating the cache, so if we use a single instance of CacheMetrics to register it as cache metrics we're safe.

Signed-off-by: Nicu Reut <nicu.reut@digitalasset.com>