Add short READMEs per Pulumi project
I had a hard time reconstructing today what some of these are about...
so perhaps nice to have a brief description right where the code is,
also for newjoiners.

Meta-goal of this PR: make sure we're all aligned about the purpose and scoping ourselves.

(Not 100% about everything here so please comment if you notice something odd.)

[static]

Signed-off-by: Martin Florian <martin.florian@digitalasset.com>