Fix reconnect logic for sv validators with bft connections (#1407)
* Fix reconnect logic for sv validators with bft connections

[ci]

Don't really see a great way of testing this but this ensure that the
validator tries to reconfigure the connections on startup
already. Otherwise you can end up in in a situation where the existing
sequencers are broken and it just waits for a connection instead of
fixing it to the current ones.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>

* .

[ci]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>

* paint the bikeshed a different color

[static]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>

---------

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>