Fix branch reference in caching definitions (#1431)
github head_ref is empty on main and only set for pull requests. So we
write the cache for `branch:` and then try to read it as
`branch:main` which is always a cache miss.

Note that main is also still not correct for PRs against the
canton-3.4 branch. But that's an orthogonal issue, I'll create a
separate issue for that.

fixes #1426 (hopefully, I can't easily test this until this is merged
to main, pretty confident it won't break it more than it already is
though)

[static]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>