Do not use rust-parallel on macos (#1464)
[static]

It looks like in newer nix versions it may work on macos
https://github.com/NixOS/nixpkgs/commit/02e74dfd6dade15eba110f0969b3f7c1896b7ca3. But
I don't have access to a macos machine to test this so I'll leave it
to the macos users to debug this if they care about parallelism. I
primarily care about CI speedups here which luckily runs on Linux.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>