Fix skipSynchronizerInitialization flag (#1495)
[ci]

So this is embarassing …

What happened is:

1. I had it all working with working tests
2. I then refactored the test and accidentally turned it into a noop.
3. I then refactored the backend and tested that the tests still work
which was not very helpful.

I guess the lesson is if you refactor tests make sure to check that
the test does not just succeed but more importantly that it can
fail. Or at least read the logs to see that it does what you expect.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>