Improve DR docs
TBH I'm not sure if step 1 is particularly practical for anyone to
do. But checking that they are on one of the two nodes can also be
done by an SV (e.g. us) or anyone else. And so far I have always seen
all parties migrated or none migrated so you don't need the manual
migrate step.

Step 2 on the other hand is fairly practical and I've recommended that
in various forms in the past with success.

[static]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>