cluster: Build cross-platform Docker images natively (#1890)
* cluster: Build cross-platform Docker images natively

This change removes hacks for building cross-platform images to run
native builds instead. This is cleaner and more reliable.

Signed-off-by: Stanislav German-Evtushenko <ginermail@gmail.com>

* cluster, images, canton: Update comment and lablel

Replace TODO by an actual comment and update label with a reference to
the actual base image

Signed-off-by: Stanislav German-Evtushenko <ginermail@gmail.com>

* scripts/check-base-images.sh: Parse Dockerfiles to get base image

This allows to set BASE_IMAGE as an argument and then use in FROM and
LABEL instead of duplicating it.

Signed-off-by: Stanislav German-Evtushenko <ginermail@gmail.com>

* scripts/check-base-images.sh: Make shellcheck happy

False-positive

Signed-off-by: Stanislav German-Evtushenko <ginermail@gmail.com>

---------

Signed-off-by: Stanislav German-Evtushenko <ginermail@gmail.com>