Fix some grafana log warnings
The sustained traffic one never got merged to main but it got added to
some clusters for testing and now keeps logging warnings.

I'm not actually completely sure about the null check in the
automation alert because I'm struggling to reproduce it but it seems
like it at least doesn't make things worse.

[static]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>