Remove config transform to bump canton domain ports (#2014)
[ci]

The split here doesn't really make sense. The one without "domain" in
the name bumps the internal API URL. The one with domain in the name
bumps the external API URL. Clearly both are domain related. And we
always set both because otherwise you get a mismatch since internal
and external are the same in integration tests.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>
Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>