Don't go through amulet rules for DsoRules_ClaimExpiredRewards
See https://github.com/DACH-NY/canton-network-internal/issues/1435
for the motivation.

Note: I deliberately did _not_ stub out the choice in
AmuletRules_ClaimExpiredRewards just yet. This means there will not be
a version bump for validators which allows us to roll out this upgrade
faster. We can do the cleanup in a later step when we have an amulet
change anyway.

[ci]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>