Upgrade lnav to 0.13.2
[static]

I've run it for a bit locally now and it seems to work a bit better
for me than the old one (fewer crashes and lockups, maybe a bit faster).

The tmpdir fix just makes sure that if tmpdir is not set it doesn't
end up being set to an empty string. Copy pasted from Canton.

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>