Use last sequencing time in catchup dashboard
Metric doesn't exist for the sequencer so this only changes it for
mediator and participant.

It's slightly awkward for the sv runbook which seems to show unix
epoch while it's onboarding. But it does now properly show the delay
at exactly every 30min that we see so overall still seems like an
improvement. We could filter out everything above one year I guess.

[static]

Signed-off-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>