Unbump resources on scratch networks
Save some money.

Part of #2943

Notably this also fixes the logic for `SEQUENCER_LOW_RESOURCES` (yes it was broken) ... and removes that flag altogether in favor of explicit values in the config. Because it seems to me that those flags and `if`s in the pulumi add more complexity than is worth it.

Tested expected behaviour via a manual scratch deploy.

[static]

Signed-off-by: Martin Florian <martin.florian@digitalasset.com>